feat: add assetId to createGAOrder
