+++ package.json
@@ -23,7 +23,7 @@
     "@babel/runtime": "^7.12.5",
     "@bit/joshk.vue-spinners-css.spinner-loader": "^1.1.5",
     "@debionetwork/pinata-ipfs": "0.0.2",
-    "@debionetwork/polkadot-provider": "0.0.13-beta.0",
+    "@debionetwork/polkadot-provider": "0.0.14",
     "@debionetwork/ui-components": "0.0.9",
     "@debionetwork/ui-icons": "0.0.3",
     "@kiltprotocol/sdk-js": "^0.21.0",
