--- src/common/lib/polkadot-provider/command/geneticAnalysisOrders.js
@@ -1,6 +1,0 @@
-export async function setGeneticAnalysisPaid(api, pair, orderId) {
-  const result = await api.tx.geneticAnalysisOrders
-    .setGeneticAnalysisOrderPaid(orderId)
-    .signAndSend(pair, { nonce: -1})
-  return result.toHuman()
-}
